\name{wheat}
\alias{wheat}
\title{wheat dataset}
\description{

Information from a collection of 599 historical CIMMYT wheat lines.  The wheat data set is from 
CIMMYT's Global Wheat Program. Historically, this program has conducted numerous international 
trials across a wide variety of wheat-producing environments. The environments represented in 
these trials were grouped into four basic target sets of environments comprising four 
main agroclimatic regions previously defined and widely used by CIMMYT's Global Wheat Breeding Program. 
The phenotypic trait considered here was the grain yield (GY) of the 599 wheat lines evaluated 
in each of these four mega-environments with two replications. 


Wheat lines were genotyped using 1447 Diversity Array Technology (DArT) generated by 
Triticarte Pty. Ltd. (Canberra, Australia; \url{http://www.triticarte.com.au}). The DArT markers 
may take on two values, denoted by their presence or absence. Markers with a minor allele frequency 
lower than 0.05 were removed, and missing genotypes were imputed with samples from the marginal 
distribution of marker genotypes, that is, \eqn{x_{ij}=Bernoulli(\hat p_j)}, where  \eqn{\hat p_j}  
is the estimated allele frequency computed from the non-missing genotypes. The number of DArT 
MMs after editting was 1279.

}

\usage{
data(wheat)
}

\format{
The data.frame \code{\link{wheat.Y}} contains the phenotypes, Matrix \code{\link{wheat.G}} is the genomic relatinship matrix calculated from the DArT markers. 
}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\references{
 Crossa, J., G. de Los Campos, P. Perez, D. Gianola, J. Burgueno, J. L. Araus, D. Makumbi, R. P. Singh, S. Dreisigacker, and J. Yan. 2010. Prediction of genetic values of quantitative traits in plant breeding using pedigree and molecular markers. Genetics. 186: 713-724.  

}
\keyword{datasets}
