\name{wheat.Y}
\alias{wheat.Y}
\title{Grain yield}
\description{
a data.frame (4792 × 4) containing the grain yield of 599 lines in two blocks in four environments, the four columns are: \code{VAR} for variety identifiers, \code{ENV} for environment identifiers, \code{block} for block identifiers and \code{y} for grain yield. 
}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}
\seealso{wheat.Y50}
\keyword{datasets}
