\name{wheat.Y50}
\alias{wheat.Y50}
\title{Grain yield}
\description{
 (400 × 5 data.frame) is a subset of \code{\link{wheat.Y}} by randomly selecting 50 wheat lines from the 599 lines. The last column of wheat.Y50 is yNA: for each of the 50 wheat lines, we randomly selected one environment as missing from the four environments, and the value of yNA for these observations are NA, the values of yNA for all other observations are the same as y.
}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}
\seealso{wheat.Y}
\keyword{datasets}
