\name{wheat.G}
\alias{wheat.G}
\alias{wheat.G50}
\title{Genomic relationship matrix for the wheat dataset}
\description{
wheat.G (matrix, 599 x 599), was calculated from DArT markers. The markers were centered and standardized to a matrix \bold{X}; G matrix was calculated as: \bold{X}\bold{X}'/p , where p is the number of columns in \bold{X}. 

}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\references{
McLaren, C. G., R. Bruskiewich, A.M. Portugal, and A.B. Cosico. 2005. The International Rice Information System. 
A platform for meta-analysis of rice crop data. \emph{Plant Physiology} \bold{139}: 637-642.
}
\keyword{datasets}
