\name{wheat.G}
\alias{wheat.G}
\alias{wheat.G50}
\title{Genomic relationship matrix for the wheat dataset}
\description{
wheat.G (matrix, 599 x 599), was calculated from DArT markers. The markers were centered and standardized to a matrix \bold{X}; G matrix was calculated as: \bold{X}\bold{X}'/p , where p is the number of columns in \bold{X}. 

}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\references{
 Crossa, J., G. de Los Campos, P. Perez, D. Gianola, J. Burgueno, J. L. Araus, D. Makumbi, R. P. Singh, S. Dreisigacker, and J. Yan. 2010. Prediction of genetic values of quantitative traits in plant breeding using pedigree and molecular markers. Genetics. 186: 713-724.  
}
\keyword{datasets}
