\name{wheat.Y}
\alias{wheat.Y}
\title{Phenotype for 599 wheat lines}
\description{
a data.frame (2396 × 3) containing the grain yield of 599 lines in four environments; the three columns are: \code{VAR} for variety identifiers, \code{ENV} for environment identifiers, and \code{y} for grain yield (each entry corresponds to an average of two plot records). 
}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}
\references{
 Crossa, J., G. de Los Campos, P. Pérez, D. Gianola, J. Burgueño, J. L. Araus, D. Makumbi, R. P. Singh, S. Dreisigacker, and J. Yan. 2010. Prediction of genetic values of quantitative traits in plant breeding using pedigree and molecular markers. Genetics. 186: 713-724.  

}
\seealso{\code{\link{wheat}}, \code{\link{wheat.G}}}
\keyword{datasets}
