\name{wheat.Y}
\alias{wheat.Y}
\title{Phenotype for 599 wheat lines}
\description{
a data.frame (2396 × 3) containing the grain yield of 599 lines in four environments; the three columns are: \code{VAR} for variety identifiers, \code{ENV} for environment identifiers, and \code{y} for grain yield (each entry corresponds to an average of two plot records). 
}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}
\seealso{\code{\link{wheat}},\code{\link{wheat.G}}}
\keyword{datasets}
