\name{wheat.Y50}
\alias{wheat.Y50}
\title{Phenotype for 50 wheat lines}
\description{
a data.frame (400 × 5), which is a subset of \code{\link{wheat.Y}} by randomly selecting 50 wheat lines from the 599 lines. The last column of \code{wheat.Y50} is \code{yNA}: for each of the 50 wheat lines, we randomly selected one environment as missing from the four environments, and the value of \code{yNA} for these observations are \code{NA}, the values of \code{yNA} for all other observations are the same as \code{y}.
}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}
\seealso{\code{\link{wheat.Y}}}
\keyword{datasets}
